# openapi.json
{"openapi":"3.1.0","info":{"title":"ll_rocket","version":"0.1.0"},"paths":{"/api/v1/auth/kakao/login":{"get":{"tags":["auth"],"summary":"Kakao Login","description":"카카오 로그인 URL 반환","operationId":"auth-kakao_login","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KakaoLoginResponse"}}}}}}},"/api/v1/auth/kakao/callback":{"get":{"tags":["auth"],"summary":"Kakao Callback","description":"카카오 로그인 콜백 처리","operationId":"auth-kakao_callback","parameters":[{"name":"code","in":"query","required":true,"schema":{"type":"string","description":"카카오 인증 코드","title":"Code"},"description":"카카오 인증 코드"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/logout":{"get":{"tags":["auth"],"summary":"Logout","operationId":"auth-logout","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Auth-Logout"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["auth"],"summary":"Get Me","operationId":"auth-get_me","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"refresh_token","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Refresh Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"KakaoLoginResponse":{"properties":{"authorization_url":{"type":"string","title":"Authorization Url"}},"type":"object","required":["authorization_url"],"title":"KakaoLoginResponse"},"UserResponse":{"properties":{"nickname":{"type":"string","title":"Nickname"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"}},"type":"object","required":["nickname"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/login/access-token"}}}}}}